refactor: add a function to generate randome strings in core/utils
