+++ backend/plugins/ae/models/connection.go
@@ -28,7 +28,7 @@ import (
 	"time"
 
 	"github.com/apache/incubator-devlake/core/errors"
-	"github.com/apache/incubator-devlake/core/plugin"
+	"github.com/apache/incubator-devlake/core/utils"
 	helper "github.com/apache/incubator-devlake/helpers/pluginhelper/api"
 )
 
@@ -36,7 +36,7 @@ type AeAppKey helper.AppKey
 
 // SetupAuthentication sets up the HTTP Request Authentication
 func (aak *AeAppKey) SetupAuthentication(req *http.Request) errors.Error {
-	nonceStr := plugin.RandLetterBytes(8)
+	nonceStr := utils.RandLetterBytes(8)
 	timestamp := fmt.Sprintf("%v", time.Now().Unix())
 	sign := signRequest(req.URL.Query(), aak.AppId, aak.SecretKey, nonceStr, timestamp)
 	req.Header.Set("x-ae-app-id", aak.AppId)
